Extract superclass LevelScreen from TestLevelScreen
It's better to do this sooner rather than later, because we'll need to
do this eventually for multiple levels.
